Added Write support and bring your own uart
This adds a couple of things that are hard to untangle into different
pull requests.

1. Adds bring your own uart

For my use case I wanted to watch the serial port from another device,
_not_ the serial port that happened to be on the esp32 that esphome was
using for logging (which I would like to keep). This makes the
configuration a little harder for the normal case, but it is more
explicit.

2. Added write support

I wanted to be able to log into my other remote device, so I added write
support

3. Added multi-support

I want to be able to have multiple of these, like a serial console
server. Only `MULTI_CONF = True` was required to make this happen, once
the `uart_id` was ready.

This also removes the Arduino'ism for serial, as we are depending on
ESPhome's uart code, which is kinda cool. Should work on any controller
supported by ESPHome and any framework with this PR.